Code review
I went through and cleaned out a lot of the excess code that was just
shot-gunned in early on. I didn't try to make it all encompassing, as
the parse will take the place of a lot of this code and the rules will
be substituted in where I've hard-coded ways to deal with certain
syntax.