DevTutorial, Preliminary ResultSet Comparison
- The DevTutorial page is up. I still need to populate it, but I believe
it will do what we want (act as a place to simply enter queries for
generated NLP). To access, use the username: "dev" and password:
"SQLTutor!!!".
- Hopefully the menubar will redirect correctly now (if you're on
localhost, it will redirect to your local home, where-as online it will
redirect to its own).
- Added preliminary resultset comparison. If the sets are unordered,
does the mentioned EXCEPT deal. Yet to implement how it will let the
user know these values that were either added or missed. If the sets
were ordered, does a simple .equals() comparison. This also does not
print which results different, just that they did.