Updated database backups, many other changes (see description)
- TutorialPage now redirects if the given link is wrong and displays an
error.
- TutorialPage now redirects if the current date/time is not within the
access date/time bounds and displays an error.
- TutorialPage now holds what schema it is showing and the options
pertaining to it (this is opposed to pulling it from the userBean -
useful in that going to a link will not allow the user to leave that
tutorial and return without using the link).
- LoginRedirect will now also record the parameters of the link the user
tried to access
- Users can now see schemas which are not currently accessible due to
their open/close date/times in the schema options (trying to access them
will redirect and display an error though).
- SchemaUploadPage now has a loading widget